feat: add replace_is_ok_with_if_let_ok assist
